requirements.txt
matplotlib==3.7.2
numpy==1.25.0
Pillow==10.0.0
PyYAML==6.0
seaborn==0.12.2
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.65.0
transformers==4.17.0
